eliminate platform-dependent casts in aspen-layer tuple and index code
Replace u32-to-usize implicit casts with explicit try_from or bounded
conversions. The tuple subsystem builds keys and manages indices — on
32-bit platforms usize is 32-bit so these were safe at runtime, but the
implicit casts hide platform dependence and defeat Tiger Style's
platform-independent-by-construction principle.

Uses try_from().unwrap_or(MAX) for capacity/construction sites where
overflow is impossible (values originate as u32), match-based get()
for bounds-safe indexing, and direct usize::MAX where only comparison
against len() occurs.

+++ crates/aspen-layer/src/index/scan.rs
@@ -21,7 +21,7 @@ impl IndexScanResult {
 
     /// Get the number of results.
     pub fn len(&self) -> u32 {
-        self.primary_keys.len().min(u32::MAX as usize) as u32
+        self.primary_keys.len().min(usize::MAX) as u32
     }
 
     /// Check if the result is empty.
